pyproject.toml
[project]
name = "iprae"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alpaca-py>=0.43.2",
    "fastapi>=0.135.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pandas>=2.3.3",
    "plotly>=6.6.0",
    "pydantic>=2.12.5",
    "requests>=2.32.5",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "streamlit>=1.55.0",
    "uvicorn>=0.42.0",
    "yfinance>=1.2.0",
]
